@@ -12,43 +12,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.1.12",
-    "@angular/cdk": "~8.1.4",
-    "@angular/common": "~9.1.12",
-    "@angular/compiler": "~9.1.12",
-    "@angular/core": "~9.1.12",
-    "@angular/forms": "~9.1.12",
-    "@angular/material": "^8.1.4",
-    "@angular/platform-browser": "~9.1.12",
-    "@angular/platform-browser-dynamic": "~9.1.12",
-    "@angular/router": "~9.1.12",
+    "@angular/animations": "~10.0.4",
+    "@angular/cdk": "^10.0.2",
+    "@angular/common": "~10.0.4",
+    "@angular/compiler": "~10.0.4",
+    "@angular/core": "~10.0.4",
+    "@angular/forms": "~10.0.4",
+    "@angular/material": "^10.0.2",
+    "@angular/platform-browser": "~10.0.4",
+    "@angular/platform-browser-dynamic": "~10.0.4",
+    "@angular/router": "~10.0.4",
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
     "hammerjs": "^2.0.8",
     "rxjs": "~6.6.0",
-    "tslib": "^1.10.0",
+    "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.11",
-    "@angular/cli": "~9.1.11",
-    "@angular/compiler-cli": "~9.1.12",
-    "@angular/language-service": "~9.1.12",
+    "@angular-devkit/build-angular": "~0.1000.3",
+    "@angular/cli": "~10.0.3",
+    "@angular/compiler-cli": "~10.0.4",
+    "@angular/language-service": "~10.0.4",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~3.3.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "nodemon": "^1.19.2",
-    "protractor": "~5.4.0",
+    "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.8.3"
+    "tslint": "~6.1.0",
+    "typescript": "~3.9.7"
   }
 }
